pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nw-algorithm-biopython"
version = "1.0.0"
description = "Needleman-Wunsch Algorithm Implementation for Sequence Alignment"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Ramo2theSky", email = "your.email@example.com"}
]
keywords = [
    "bioinformatics",
    "sequence-alignment",
    "needleman-wunsch",
    "pairwise-alignment",
    "dna-sequence"
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]
requires-python = ">=3.8"
dependencies = [
    "biopython>=1.81",
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "matplotlib>=3.4.0",
    "seaborn>=0.11.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.12",
    "black>=21.0",
    "flake8>=3.9",
]
docs = [
    "sphinx>=4.0",
    "sphinx-rtd-theme>=1.0",
]

[project.urls]
Homepage = "https://github.com/Ramo2theSky/NeedlemanWunsch-Sequence-Aligner"
Documentation = "https://github.com/Ramo2theSky/NeedlemanWunsch-Sequence-Aligner/tree/main/docs"
Repository = "https://github.com/Ramo2theSky/NeedlemanWunsch-Sequence-Aligner.git"
"Bug Tracker" = "https://github.com/Ramo2theSky/NeedlemanWunsch-Sequence-Aligner/issues"
